UI: Fix dark-mode renders for code blocks
This seeks to improve the rendering of code snippets in dark mode. Notably, blocks nested inside of admonitions were hard to read in dark mode. This PR swaps out the background color value to use one of Dracula's (our syntax highlighter's theme) purple values that closely matches our admonition color, while providing good contrast and ease of reading for the bright colors of the highlighter. Further, the hardcoding of text colors has been removed; it would appear that Dracula's YAML highlighting in dark mode is the only thing that doesn't play well with our setup. But, this isn't a blocker.