Docs: Add PostgreSQL getting started (old workflow)
This is a composite commit containing a series of WIPs over the past 72 hours or so. As the new CLI is not yet in production, I wanted to write this guide using the old workflow first and have a fallback point should anything change or not make its way to production. The next commit in this branch will reflect the CLI as of 23 May 2024 🤙