Docs: Add note for for local dev
Some browsers — notably Brave — have privacy settings which block the console from accessing the engine instance on a user's machine. This just alerts them to the fact that it's like an issue on their end and to allowlist console.hasura.io in whatever fashion makes the most sense for them.