Docs: Add Explore an API page
In order to meet GA-readiness requirements, this PR adds an Explore an API page to the getting started flow. It attempts to provide a high-level overview that satisfies the requirements in our RFC and introduces a user to the Console so they can explore all that it has to offer.