Docs: Update snippet instructions
Updates the snippet instructions to not be prefixed with "From the root of your project..." where appropriate.

Importantly, the snippet instructions for many blocks retain the prefix. We've aimed to make these copy-pasta, so it's important to direct a user _back_ to the project's root in many cases so that the paths passed to various flags are accurate and don't result in the CLI returning errors.