Docs: Update Getting Started to match new DX
Updates the getting started to match the new DX from the CLI. Changes are probably too numerous to list out. However, the broad strokes are that the majority of what we had has been drastically simiplified. Additionally, where primitives can still be used, we've attempted to call them out for grokking.