Sunday, sunday 🎵
Revises the entire GS workflow to reflect the new DX. In reality, this was a quality check and there was a lot of work needed to make this accurate. However, after building a local project with all the availalbe connectors + deploying it to DDN, I feel confident about the efficacy of these updates.