Update CodeStep component
Updates the CodeStep component to use BrowserOnly and MutationObserver when rendering to prevent hydration issues.

As Sean rightly pointed out, BrowserOnly is a helpful addition. However, when wrapping the CodeStep component in this, it caused side effects in the `GuideDocs.tsx` file. Namely, there were no nodes present when the `useEffect` hook ran, so there were no numbers generated for each step. The addition of the [MutationObserver](https://developer.mozilla.org/en-US/docs/Web/API/MutationObserver) allows us to look for changes in the tree and trigger the logic when there's actually content on the page.