Docs: Add information about env vars
Adds more information about cloud env vars.

A few folks have expressed some confusion as to what is required and what isn't. As this file is generated at this point in the deployment process, we've added an admonition here to alert users to the environment variables that will be created for each conncetor and which are required / will be overwritten at runtime.