Add hashing and validation recipes
Adds to recipes for hashing and validating credentials. I started this as a single recipe, but realized there's two different use cases here

Notably, we're introducing a mutation for the first time and I have to wonder: should we be more specific by adding in a mock GraphQL mutation the user should be calling? I think not, as not all data sources support point-mutations yet. But, something to consider.