Docs: Update Auth0 guide
Updates the Auth0 guide to be clearer about testing.

In the first version of this document, we simply recommended users log their JWT via the client application which uses Auth0. In a recent addition,
it was recommended that users reach for the API Authentication Debugger provided by Auth0. However, there were reports of custom claims failing to
appear. We've relegated this option to an admonition, brought back the initial guidance, and bolstered it by providing a new repository wherein we
can add examples of client applications for various providers and allow users to verify their JWT generation on their own.