Docs: Stub out new directory and fix refs
Stubs out a new directory and fixes any refs that are now gone.

It should be noted that this massively affected two areas:
* configuration refernces for PostgreSQL, MongoDB, and ClickHouse.
* partly related to that: "basic" reference material for native operations on these connectors.

We should consider — as part of this task — stubbing out the new `Reference` section and moving
in these files so they're not lost. We can simplify what's included, but defnitely don't want to
lose valuable information that users need.