Docs: Stub out pages
Stubs out pages for the new directory.

Importantly, this leaves out the "Build your own connector" tutorial for...reasons.

Notably, the ndc-test runner needs to be updated and our underlying documentation for various
SDKs + the connector-building process need to be updated.

After spending some time this week with the TS SDK, I'm of the opinion that docs is not the place
for this (these) tutorial(s). Instead, each SDK should publish a guide for working with it...
even then, the connector authoring experience will never be very linear and tutorial-friendly.