Docs: Add ClickHouse reference docs
Adds ClickHouse reference docs.

As we're dumping the `connectors` directory, the config files that we planned to move into `reference` need a home; this needs
to be stubbed out now. This section will take some iteration but, largely, this material is moved over 1:1 to what was present
in the `connectors` directory. The main augmentation is the inclusion of index and overview pages that will certainly expand
as we add in the CLI docs along with metadata objects.