Update queries
Updates queries to include a new `Basics` page (which is effectively the overview) and also adds aggregation docs.

The latter is from another ticket we discovered in the backlog. To note: we're deviating from what we did in v2 in the
sense of we had GraphQL API docs for each individual data source (or most of them). Here, we're unifying things and
alerting users on the index of each section that they'll need to consult their connector's docs to see what is supported.

This means we should include a feature matrix with various query, mutation, and subscription types on the index page of each
connector's reference docs.