Stub out modified directory
Stubs out the modified directory by creating the shape we have in the restructure doc.

Notably, there's refs that I've had to modify in other places that now simply link to the overview page
for the directory. Additionally, I moved all the biz logic recipes into the `Tutorial` subdirectory and
placed them after two very basics tutorials: adding custom biz logic (i.e., adding a lambda connector +
returning data from a function in the API) and extending a model (i.e., creating a relationship + connecting
it the function to a model). The latter of which may have a name change to make more sense to users.