Update Basics page
Updates Basics page to follow a similar format to all restructured top-level overviews.

Notably, we're introducing the concept of business logic and our specific definition for what it is.
We address questions a user may have about what we deem business logic, why they'd want to incorporate it
into their API, how it works, where it's hosted, and which languages are supported.