Add how-to for error handling
Adds a how-to for error handling using various SDK-specific classes.

There's variability in how connectors implement this according to their language,
but all use the ndc spec for guidance on which errors can be returned.

Additionally, we let users know bset practices (not passing the entire stack trace)
and how to access the otel collector via the console.