Docs: Add standalone guide
Adds a standalone guide. As it is, this guide should work. However, it's intentionaly left a little vague to put the onus on the user who wishes to self-deploy. I'm going to spend some time going through the steps and see if there's at least one or two places we can offer more guidance so as to make this a bit more welcoming.