Docs: Add how-to-build guide for Trino
Adds a how-to-build guide in the style of all the others during this sprint.

Notably, as Trino is a query engine (capable of connecting to multiple distrubted services), we chose the simplest option: Postgres for what to connect to. Arguably, it may be of value to connect multiple sources for a more authentic experience, but this is step 1 for the MVP.