CI: Add CanonicalUrl component
Adds a `CanonicalUrl` component that allows us to explicitly define a `canonicalUrl` in frontmatter for any page we wish.

This allows an author to add `canonicalUrl: https://promptql.io/docs/<page_path>` for any content which is duplicated and we want to promote SEO for the PromptQL version.

For pages that _don't_ contain `canonicalUrl` frontmatter, Docusaurus will continue to genreate the canonical URL automatically.

Content changes will be incorporated in a follow-up PR that branches off this feature branch.